Add Github secrets handling to actions
This update allows for the inclusion of Github secrets in the actions configuration. The secrets can now be passed to Shell - no longer limited only to visible variables. The new setup also introduces a 'secrets' field in the 'action.yml' file, and the 'test_action.yml' file now displays its implementation.