Refactor codebase for Docker compatibility and add new workflows
The update modifies the main Python script to work with Docker and alters the action.yml file to utilize a Docker image. Test files and workflows have been removed to streamline the codebase. It also includes the addition of two new GitHub workflows: lint.yml for code linting, and integration.yml for integration testing along with a new Dockerfile for containerization.