Switch from environment variable to command line argument for files
The environment variable `INPUT_FILES` in `action.yml` and `src/main.py` has been replaced with a command line argument `INPUT_FILES_ARG`. This has been done to be more flexible with the passing of arguments. As a result, the Dockerfile had to be modified to use this new argument in the CMD instruction.